[Fix][Windows] Prevent deadlock in FBuildWorker on shutdown when it's sent a WM_QUIT by another application
